+++ tests/testthat/test-style_tbl_compact.R
@@ -2,28 +2,28 @@ test_that("no errors with typical use", {
   expect_error(
     head(trial) %>%
       gt::gt() %>%
-      compact_table(),
+      style_tbl_compact(),
     NA
   )
 
   expect_error(
     head(trial) %>%
       flextable::flextable() %>%
-      compact_table(),
+      style_tbl_compact(),
     NA
   )
 
   expect_error(
     head(trial) %>%
       huxtable::huxtable() %>%
-      compact_table(),
+      style_tbl_compact(),
     NA
   )
 
   expect_error(
     head(trial) %>%
       knitr::kable(format = "html") %>%
-      compact_table(),
+      style_tbl_compact(),
     NA
   )
 })