+++ NEWS.md
@@ -1,6 +1,6 @@
 # bstfun 0.1.3
 
-* Added new function `compact_table()` that makes the compact styling available in `gtsummary::theme_gtsummary_compact()` to any {gt}, {flextable}, {huxtable}, or `knitr::kable()` table.
+* Added new function `style_tbl_compact()` that makes the compact styling available in `gtsummary::theme_gtsummary_compact()` to any {gt}, {flextable}, {huxtable}, or `knitr::kable()` table.
 
 # bstfun 0.1.2