+++ NEWS.md
@@ -1,4 +1,4 @@
-# bstfun (development version)
+# bstfun 0.1.4
 
 * Added `hpcc_get_arg()` and `hpcc_get_seq_number()` functions to aid working with the high performance computing cluster. (#19)